fixes for Windows
- added detection and proper handling for sleep() + unistd.h / Sleep()
  to have a cross-platform yield implementation (needed with the GPU
  waiting hack with ECC ON);
- made setting a few GPU force buffer values to NaN cross-platform
- disabled CUDA <= 4.0 on Windows with MSVC;
- -fPIC is now only used when it is supported and it is added only to
  NVCC flags;
- don't set automatically the nvcc host compiler with MSVC because nvcc
  already does this behind the scenes.

+++ src/gmxlib/bondfree.c
@@ -3149,7 +3149,7 @@ void init_bonded_thread_force_reduction(t_forcerec *fr,
      * Minimum force block reduction size is 2^6=64.
      */
     fr->red_ashift = 6;
-    while (fr->natoms_force > MAX_BLOCK_BITS*(1U<<fr->red_ashift))
+    while (fr->natoms_force > (int)(MAX_BLOCK_BITS*(1U<<fr->red_ashift)))
     {
         fr->red_ashift++;
     }
